# frontend/openapi/schemas/be-schema.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/be-nest-stocks/api/stocks/{symbol}": {
            "get": {
                "operationId": "StocksController_getStock",
                "parameters": [
                    {
                        "name": "symbol",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/be-nest-stocks/api/stocks/update": {
            "post": {
                "operationId": "StocksController_updateStocksList",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateStockListDto"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            }
        },
        "/be-nest-stocks/api/fmp/{symbol}/update-statements": {
            "post": {
                "operationId": "FmpController_fetchAndSaveStatements",
                "parameters": [
                    {
                        "name": "symbol",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
...
        "/be-nest-stocks/api/fmp/{symbol}/update-analyst-ratings": {
            "post": {
                "operationId": "FmpController_fetchAndSaveRarings",
                "parameters": [
                    {
                        "name": "symbol",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/be-nest-stocks/api/fmp/{symbol}/update-outlook": {
            "post": {
                "operationId": "FmpController_fetchAndSaveOutlook",
                "parameters": [
                    {
                        "name": "symbol",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                }
            }
        },
        "/be-nest-stocks/api/fmp/{symbol}/combine": {
            "get": {
                "operationId": "FmpController_getCombinedData",
                "parameters": [
                    {
                        "name": "symbol",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
...
        "/be-nest-stocks/api/user/register": {
            "post": {
                "operationId": "UsersController_register",
                "summary": "Register a new user",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": ["users"]
            }
        },
        "/be-nest-stocks/api/user/login": {
            "post": {
                "operationId": "UsersController_login",
                "summary": "Login a user",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": ["users"]
            }
        },
        "/be-nest-stocks/api/user/portfolio": {
            "put": {
                "operationId": "UsersController_addToPortfolio",
                "summary": "Add a stock to portfolio list",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePortfolioDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": ["users"],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "operationId": "UsersController_removeFromPortfolio",
                "summary": "Remove a stock from portfolio list",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePortfolioDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": ["users"],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "get": {
                "operationId": "UsersController_getPortfolioList",
                "summary": "Get portfolio list",
                "parameters": [
                    {
                        "name": "email",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": ["users"],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/be-nest-stocks/api/user/consider": {
            "put": {
                "operationId": "UsersController_addToConsider",
                "summary": "Add a stock to consider list",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateConsiderDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": ["users"],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "operationId": "UsersController_removeFromConsider",
                "summary": "Remove a stock from consider list",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateConsiderDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": ["users"],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "get": {
                "operationId": "UsersController_getConsiderList",
                "summary": "Get consider list",
                "parameters": [
                    {
                        "name": "email",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": ["users"],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "info": {
        "title": "Stock Screener API",
        "description": "API for stock screener application",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
            "UpdateStockListDto": {
                "type": "object",
                "properties": {}
            },
...
            "RegisterDto": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "user@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "securepassword"
                    }
                },
                "required": ["email", "password"]
            },
            "LoginDto": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "user@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "securepassword"
                    }
                },
                "required": ["email", "password"]
            },
            "UpdatePortfolioDto": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "user@example.com"
                    },
                    "stockId": {
                        "type": "string",
                        "example": "6652f9ee06baf75ceb6c219e"
                    }
                },
                "required": ["email", "stockId"]
            },
            "UpdateConsiderDto": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "user@example.com"
                    },
                    "stockId": {
                        "type": "string",
                        "example": "6652f9ee06baf75ceb6c219e"
                    }
                },
                "required": ["email", "stockId"]
            }
        }
    }
}